move unit tests from "specs" dir to the corresponding directories
