+++ .eslintrc
@@ -40,6 +40,7 @@
     "node": true
   },
   "globals": {
-    "__impl__": true
+    "describe": true,
+    "it": true
   }
 }
