+++ fixtures/dependencies-fixture.json
