+++ fixtures/jsdoc/endWithTwoStars.js
@@ -5,5 +5,5 @@
  * <div>
  *  <StoreList data={[{summary:"Item", details:"Convallis qui non phasellus nullam congue et pellentesque eros"}]} />
  * </div>
-**/
-export default class Accordion extends React.Component { }
+ * */
+export default class Accordion extends React.Component {}
