+++ fixtures/jsdoc/functionDeclaration.js
@@ -2,20 +2,20 @@
  * Adds two numbers.
  *
  * @name add
- * @category Math
+ * @static
+ * @public
  * @param {number} a The first number in an addition.
  * @param {number} b The second number in an addition.
  * @returns {number} Returns the total.
  * @example
- *
- * _.add(6, 4);
- * // => 10
+ * //-
+ * // Adds two numbers
+ * //-
+ * add(2, 3);
+ * //=>
+ * // 5
+ * //=>
  */
-var add = function(a, b) {
-  return a+b;
-};
-
-module.exports = {
-  add,
-};
-
+function add(a, b) {
+  return a + b;
+}
