+++ fixtures/jsdoc/methodDeclaration.js
