+++ package.json
@@ -9,24 +9,24 @@
     "precommit": "lint-staged",
     "flow": "flow; test $? -eq 0 -o $? -eq 2",
     "flow-coverage": "flow-coverage-report -i \"./src/**/*.js\"",
-    "lint": "eslint \"{src,specs,e2e}/**/*.js\" && flow check",
-    "lint:table": "eslint \"{src,specs,e2e}/**/*.js\" --format table",
-    "lint:html": "eslint \"{src,specs,e2e}/**/*.js\" --format html -o eslint-report.html",
-    "lint-circle": "eslint \"{src,specs,e2e}/**/*.js\" --ignore-path .eslintignore-circle --format junit -o junit/eslint-results.xml",
-    "lint-circle-debug": "eslint \"{src,specs,e2e}/**/*.js\" --ignore-path .eslintignore-circle",
-    "format": "prettier-eslint \"{src,specs,e2e}/**/*.js\" --write",
-    "test": "mocha --require babel-core/register './specs/**/*.spec.js'",
-    "test:appveyor": "mocha --require babel-core/register --reporter mocha-appveyor-reporter ./specs/**/*.spec.js",
-    "test-circle": "mocha --require babel-core/register --reporter mocha-circleci-reporter './specs/**/*.spec.js'",
+    "lint": "eslint \"{src,e2e}/**/*.js\" && flow check",
+    "lint:table": "eslint \"{src,e2e}/**/*.js\" --format table",
+    "lint:html": "eslint \"{src,e2e}/**/*.js\" --format html -o eslint-report.html",
+    "lint-circle": "eslint \"{src,e2e}/**/*.js\" --ignore-path .eslintignore-circle --format junit -o junit/eslint-results.xml",
+    "lint-circle-debug": "eslint \"{src,e2e}/**/*.js\" --ignore-path .eslintignore-circle",
+    "format": "prettier-eslint \"{src,e2e}/**/*.js\" --write",
+    "test": "mocha --require babel-core/register './src/**/*.spec.js'",
+    "test:appveyor": "mocha --require babel-core/register --reporter mocha-appveyor-reporter ./src/**/*.spec.js",
+    "test-circle": "mocha --require babel-core/register --reporter mocha-circleci-reporter './src/**/*.spec.js'",
     "mocha-circleci": "mocha --require babel-core/register --reporter mocha-circleci-reporter",
     "e2e-test": "mocha --require babel-core/register ./e2e/**/*.e2e.js",
     "e2e-test:appveyor": "mocha --require babel-core/register --reporter mocha-appveyor-reporter ./e2e/**/*.e2e.js",
     "e2e-test-circle": "mocha --require babel-core/register --reporter mocha-circleci-reporter ./e2e/**/*.e2e.js",
     "clean": "rm -rf dist",
-    "build": "babel src specs -d dist",
+    "build": "babel src -d dist",
     "pkg": "pkg . --out-path releases/",
-    "watch": "babel --watch bin src specs -d dist --source-maps",
-    "build-debug": "babel src specs -d dist --source-maps",
+    "watch": "babel --watch bin src -d dist --source-maps",
+    "build-debug": "babel src -d dist --source-maps",
     "prepublishOnly": "npm run clean && npm run build"
   },
   "bin": {
