--- specs/.eslintrc
@@ -1,12 +1,0 @@
-{
-  "rules": {
-    "no-unused-expressions": [0]
-  },
-  "env": {
-    "mocha": true
-  },
-  "globals": {
-    "describe": true,
-    "it": true
-  }
-}
