+++ src/api/consumer/lib/get-scope-bit.spec.js
