+++ src/api/consumer/lib/list-scope.spec.js
