+++ src/bit-id/bitId.spec.js
