+++ src/consumer/component/consumer-component.spec.js
