+++ src/npm-client/npm-client.spec.js
