+++ src/search/indexer.spec.js
