+++ src/utils/buffer/replace-buffer-non-recursive.spec.js
